Commit committed...

+++ HW3/src/system/Continuation.java
@@ -5,11 +5,11 @@ package system;
  */
 public class Continuation{
 
-    public final int closureId;
+    public final long closureId;
     public final int offset;
     public final Object argument;
 
-    public Continuation(int closureId, int offset, Object argument) {
+    public Continuation(long closureId, int offset, Object argument) {
         this.closureId = closureId;
         this.offset = offset;
         this.argument = argument;
